Reorganize DashboardPicture and Avatar components
