+++ src/components/Avatar/Avatar.scss
@@ -1,13 +1,3 @@
-.picture-wrapper {
-  display: flex;
-  align-items: center;
-}
-
-.dash-heading {
-  width: 100px;
-  text-align: left;
-}
-
 .image-input {
   opacity: 0;
   position: absolute;
