+++ src/components/Avatar/index.js
@@ -1,0 +1,3 @@
+import Avatar from './Avatar';
+
+export default Avatar;
