+++ src/components/DashboardPicture/DashboardPicture.scss
@@ -1,0 +1,9 @@
+.picture-wrapper {
+  display: flex;
+  align-items: center;
+}
+
+.dash-heading {
+  width: 100px;
+  text-align: left;
+}
