+++ src/components/DashboardPicture/index.js
@@ -1,0 +1,3 @@
+import DashboardPicture from './DashboardPicture';
+
+export default DashboardPicture;
