--- src/components/UserPicture/index.js
@@ -1,3 +1,0 @@
-import UserPicture from './UserPicture';
-
-export default UserPicture;
